Moving benchmark utilities to benchmarks package (#53)
