+++ go.mod
@@ -3,8 +3,8 @@ module github.com/xybor-x/xylog
 go 1.18
 
 require (
-	github.com/xybor-x/xycond v1.0.0-beta.1
-	github.com/xybor-x/xyerror v0.1.0-alpha
+	github.com/xybor-x/xycond v1.0.0
+	github.com/xybor-x/xyerror v1.0.0
 	github.com/xybor-x/xylock v0.0.1
 )
